requirements.txt
pandas
numpy
matplotlib
seaborn
python-dotenv
groq
